+++ SRC/sgglse.f
@@ -350,7 +350,7 @@
       CALL SORMRQ( 'Left', 'Transpose', N, 1, P, B, LDB, WORK( 1 ),
      $             X,
      $             N, WORK( P+MN+1 ), LWORK-P-MN, INFO )
-      WORK( 1 ) = P + MN + MAX( LOPT, INT( WORK( P+MN+1 ) ) )
+      WORK( 1 ) = REAL( P + MN + MAX( LOPT, INT( WORK( P+MN+1 ) ) ) )
 *
       RETURN
 *
