+++ SRC/stgex2.f
@@ -289,7 +289,7 @@
       M = N1 + N2
       IF( LWORK.LT.MAX( N*M, M*M*2 ) ) THEN
          INFO = -16
-         WORK( 1 ) = MAX( N*M, M*M*2 )
+         WORK( 1 ) = REAL( MAX( N*M, M*M*2 ) )
          RETURN
       END IF
 *
