+++ SRC/strsyl3.f
@@ -244,8 +244,8 @@
       IWORK( 1 ) = NBA + NBB + 2
       IF( LQUERY ) THEN
          LDSWORK = 2
-         SWORK( 1, 1 ) = MAX( NBA, NBB )
-         SWORK( 2, 1 ) = 2 * NBB + NBA
+         SWORK( 1, 1 ) = REAL( MAX( NBA, NBB ) )
+         SWORK( 2, 1 ) = REAL( 2 * NBB + NBA )
       END IF
 *
 *     Test the input arguments
@@ -1175,8 +1175,8 @@
 *        form (1/SCALE)*X if SCALE is REAL. Set SCALE to zero and give up.
 *
          IWORK(1) = NBA + NBB + 2
-         SWORK(1,1) = MAX( NBA, NBB )
-         SWORK(2,1) = 2 * NBB + NBA
+         SWORK(1,1) = REAL( MAX( NBA, NBB ) )
+         SWORK(2,1) = REAL( 2 * NBB + NBA )
          RETURN
       END IF
 *
@@ -1239,8 +1239,8 @@
 *     Restore workspace dimensions
 *
       IWORK(1) = NBA + NBB + 2
-      SWORK(1,1) = MAX( NBA, NBB )
-      SWORK(2,1) = 2 * NBB + NBA
+      SWORK(1,1) = REAL( MAX( NBA, NBB ) )
+      SWORK(2,1) = REAL( 2 * NBB + NBA )
 *
       RETURN
 *
